Change the emacsGcc attribute to emacsNativeComp
In upstream nixpkgs PR https://github.com/NixOS/nixpkgs/pull/168076
native compilation support was introduced to the nixpkgs attribute.
This is a better naming more in line with upstream, so we should also
adopt it.

I have aliased emacsGcc to emacsNativeComp for backwards compatibility
with a trace notice informing users about the change.

Additionally we will now default to the upstream native comp derivation if it exists.