overlays: use a better and clearer way to compose overlays
The old way of composing overlays addes an unnecessay overlay, which
breaks the packages attribute of the flake. To reduce the impact on
users, I make a workaround[1] for that, which has been merged. Now I
figure out the real fix, which is to change (_: super) to (_: {}).

Additionally, the old way is a bit confusing.

This patch fixes both issues by using a better and clearer way to
compose overlays.

Since the workaround is no longer needed after this patch, I revert
that in the next commit.

[1]: https://github.com/nix-community/emacs-overlay/pull/256