Remove exwm repo
Exwm is in elpa and the update cadence doesn't motivate that we have the special handling around any more.

This removal feels quite funny given that this repository started it's
life as an overlay precisely to get bleeding edge exwm versions, and then grew to other packages from there.

So long exwm repo!