recipes-archive-melpa: fix incorrect source hashes
This is basically https://github.com/NixOS/nixpkgs/pull/309270 with
two fixes excluded.

The fix for smartparens is excluded because it has already been
fixed[1] here.

The fix for flycheck-cfn is also excluded because it has been bumped
to 20240512.2341 in 4fe6cc37d9fee05e46aed5cc94ed145defb72380.
However, the new hash is still incorrect.  Its fix is in the next
commit.

[1]: https://github.com/nix-community/emacs-overlay/pull/400